+++ packages/toml/package.json
@@ -5,14 +5,16 @@
     "pkg/covector_toml_bg.wasm",
     "pkg/covector_toml.js",
     "pkg/covector_toml.d.ts",
-    "index.js",
-    "index.d.ts"
+    "dist/index.js",
+    "dist/index.d.ts"
   ],
-  "main": "index.js",
-  "types": "index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "build": "wasm-pack build --release --no-pack --target nodejs && rimraf pkg/.gitignore",
-    "build:debug": "wasm-pack build --dev --no-pack --target nodejs && rimraf pkg/.gitignore",
+    "build": "wasm-pack build --release --no-pack --target nodejs && rimraf pkg/.gitignore && npm run build:ts",
+    "build:debug": "wasm-pack build --dev --no-pack --target nodejs && rimraf pkg/.gitignore && npm run build:ts",
+    "build:ts": "tsc -b",
+    "clean": "rimraf dist tsconfig.tsbuildinfo node_modules",
     "pretest": "npm run build",
     "test": "jest"
   },
