+++ tsconfig.packages.json
@@ -7,6 +7,7 @@
     { "path": "./packages/changelog" },
     { "path": "./packages/assemble" },
     { "path": "./packages/apply" },
+    { "path": "./packages/toml" },
     { "path": "./packages/covector" }
   ]
 }
